Fixed small bugs at line 87
The timestamp is used in the name of each file, and this could be confused with the year information for the function "get_file_by_year". 

For example, if the timestamp is Nov 24, 2020, that is "20201124" in the name, the model would consider this as the year 2011 data. 

To fix it with future use, I simply compared it with "f[-7:-3]", which is the year of outputs. 

Lei